+++ docs_snippets/lib/usages/widgets/popover_menu.dart
@@ -43,7 +43,6 @@ final popoverMenu = FPopoverMenu(
   // {@category "Core"}
   style: const .delta(maxWidth: 250),
   divider: .full,
-  hover: true,
   menu: [
     .group(
       children: [
@@ -97,7 +96,6 @@ final popoverMenuTiles = FPopoverMenu.tiles(
   // {@category "Core"}
   style: const .delta(maxWidth: 250),
   divider: .full,
-  hover: true,
   menu: [
     .group(
       children: [